% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_DATASET}
\alias{Example_DATASET}
\title{Aggregated data of two schools an educational survey}
\format{
A data frame with 60 rows and 2 variables:
\describe{
  \item{group}{Participant's group, A or B.}
  \item{score}{The participant's score in hypothetical task.}
}
}
\source{
<https://www.github.com/mvuorre/exampleRPackage>
}
\usage{
Example_DATASET
}
\description{
A data set with the scores of two schools
}
\keyword{datasets}
