% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{Example Function.}
\usage{
example_function(a, ..., delta = 1)
}
\arguments{
\item{a}{numeric argument}

\item{...}{Not, used- force later arguments to bind by name}

\item{delta}{numeric amount to add}
}
\value{
a + delta
}
\description{
This function is an example.
}
\examples{

example_function(3, delta = 2)  # should be 5

}
