% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Scores of Group A and Group B}
\format{
A data frame with 60 rows and 2 variables:
\describe{
\item{group}{Participant's group, A or B.}
\item{score}{The participant's score in hypothetical task.}
}
}
\source{
\url{https://www.github.com/mvuorre/exampleRPackage}
}
\usage{
exampleData
}
\description{
A data set with the scores of two groups.
}
\keyword{datasets}
