% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_help.R
\docType{package}
\name{ExampleRPackage}
\alias{ExampleRPackage}
\title{ExampleRPackage: Evocative Package Title (Title Case)}
\description{
Short description of why one should use this package.
Used by CRAN, so check their rules on "Writing R Extensions"
\url{https://cran.r-project.org/doc/manuals/r-release/R-exts.html}.
}
