% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func1.R
\name{func1}
\alias{func1}
\title{Title}
\usage{
func1(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Twice the value of the input.
}
\description{
A function to double the value of the input.
}
\examples{
func(c(1,2))
}
