% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_in_days.R
\name{get_age_in_days}
\alias{get_age_in_days}
\title{Calculate age in days}
\usage{
get_age_in_days(age_years)
}
\arguments{
\item{age_years}{A numeric scalar, representing the age of an individual
in years.}
}
\value{
A numeric scalar representing the age of the individual in days.
}
\description{
Convert the age of an individual in years to the corresponding age in days,
by multiplying with 365.25 and rounding to the nearest integer.
}
\examples{
get_age_in_days(2)

}
\author{
Charlotte Soneson
}
