% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num_not_miss.R
\name{Num_not_miss}
\alias{Num_not_miss}
\title{Calculate number of missingness}
\usage{
Num_not_miss(x)
}
\arguments{
\item{x}{Input can be a vector or a column of data frame}
}
\value{
a numeric number
}
\description{
This function provides a quick estimate of how many missing values found.
}
\examples{
Num_not_miss(c(1,2,NA,NA))
}
