% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Say hello}
\usage{
say_hello(name)
}
\arguments{
\item{name}{Character scalar giving the name
of the person to say hello to.}
}
\value{
Doesn't return anything, prints out a message.
}
\description{
Say hello
}
\examples{
say_hello("Jane")

}
