% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Say hello}
\usage{
say_hello(name)
}
\arguments{
\item{name}{A character scalar, representing the name of the person to say
hello to. The hello message will be generated using the
\code{\link[base]{message}} function}
}
\value{
Returns \code{NULL}. A message with a greeting is generated.
}
\description{
Say hello
}
\examples{
say_hello("Jane")

}
\author{
Jonas Oefelein
}
